requirements.txt
torch==2.9.1
ultralytics==8.3.241
opencv-python==4.12.0.88
numpy==2.2.6
pandas==2.3.3
tqdm>=4.66.0
pyyaml>=6.0.0
scikit-image==0.26.0
rich>=14.2.0
timm==1.0.22
